+++ src/filter/filters/crop/gui/filtergui_crop.cpp
@@ -15,13 +15,13 @@ filtergui_crop_c::filtergui_crop_c(abstract_filter_c *const filter)
     {
         auto *xPos = new filtergui_spinbox_s;
         xPos->get_value = [=]{return filter->parameter(filter_crop_c::PARAM_X);};
-        xPos->set_value = [=](const double value){filter->set_parameter(filter_crop_c::PARAM_X, value);};
+        xPos->set_value = [=](const int value){filter->set_parameter(filter_crop_c::PARAM_X, value);};
         xPos->maxValue = MAX_CAPTURE_WIDTH;
         xPos->minValue = 0;
 
         auto *yPos = new filtergui_spinbox_s;
         yPos->get_value = [=]{return filter->parameter(filter_crop_c::PARAM_Y);};
-        yPos->set_value = [=](const double value){filter->set_parameter(filter_crop_c::PARAM_Y, value);};
+        yPos->set_value = [=](const int value){filter->set_parameter(filter_crop_c::PARAM_Y, value);};
         yPos->maxValue = MAX_CAPTURE_HEIGHT;
         yPos->minValue = 0;
 
@@ -31,13 +31,13 @@ filtergui_crop_c::filtergui_crop_c(abstract_filter_c *const filter)
     {
         auto *width = new filtergui_spinbox_s;
         width->get_value = [=]{return filter->parameter(filter_crop_c::PARAM_WIDTH);};
-        width->set_value = [=](const double value){filter->set_parameter(filter_crop_c::PARAM_WIDTH, value);};
+        width->set_value = [=](const int value){filter->set_parameter(filter_crop_c::PARAM_WIDTH, value);};
         width->maxValue = MAX_CAPTURE_WIDTH;
         width->minValue = 1;
 
         auto *height = new filtergui_spinbox_s;
         height->get_value = [=]{return filter->parameter(filter_crop_c::PARAM_HEIGHT);};
-        height->set_value = [=](const double value){filter->set_parameter(filter_crop_c::PARAM_HEIGHT, value);};
+        height->set_value = [=](const int value){filter->set_parameter(filter_crop_c::PARAM_HEIGHT, value);};
         height->maxValue = MAX_CAPTURE_HEIGHT;
         height->minValue = 1;
 
@@ -47,7 +47,7 @@ filtergui_crop_c::filtergui_crop_c(abstract_filter_c *const filter)
     {
         auto *scaler = new filtergui_combobox_s;
         scaler->get_value = [=]{return filter->parameter(filter_crop_c::PARAM_SCALER);};
-        scaler->set_value = [=](const double value){filter->set_parameter(filter_crop_c::PARAM_SCALER, value);};
+        scaler->set_value = [=](const int value){filter->set_parameter(filter_crop_c::PARAM_SCALER, value);};
         scaler->items = {"Linear", "Nearest", "None"};
 
         this->guiFields.push_back({"Scaler", {scaler}});
