+++ src/filter/filters/decimate/gui/filtergui_decimate.cpp
@@ -15,7 +15,7 @@ filtergui_decimate_c::filtergui_decimate_c(abstract_filter_c *const filter)
         auto *const factor = new filtergui_spinbox_s;
 
         factor->get_value = [=]{return filter->parameter(filter_decimate_c::PARAM_FACTOR);};
-        factor->set_value = [=](const double value){filter->set_parameter(filter_decimate_c::PARAM_FACTOR, std::pow(2, value));};
+        factor->set_value = [=](const int value){filter->set_parameter(filter_decimate_c::PARAM_FACTOR, std::pow(2, value));};
         factor->minValue = 1;
         factor->maxValue = 4;
 
@@ -26,7 +26,7 @@ filtergui_decimate_c::filtergui_decimate_c(abstract_filter_c *const filter)
         auto *const sampler = new filtergui_combobox_s;
 
         sampler->get_value = [=]{return filter->parameter(filter_decimate_c::PARAM_TYPE);};
-        sampler->set_value = [=](const double value){filter->set_parameter(filter_decimate_c::PARAM_TYPE, value);};
+        sampler->set_value = [=](const int value){filter->set_parameter(filter_decimate_c::PARAM_TYPE, value);};
         sampler->items = {"Nearest", "Average"};
 
         this->guiFields.push_back({"Sampler", {sampler}});
